+++ selfdrive/controls/lib/lead_mpc_lib/SConscript
@@ -43,6 +43,6 @@ if GetOption('mpc_generate'):
     env.Command(generated_c + generated_h, generator, cmd)
 
 
-mpc_files = ["longitudinal_mpc.c"] + generated_c
+mpc_files = ["lead_mpc.c"] + generated_c
 env.SharedLibrary('mpc1', mpc_files, LIBS=['m', 'qpoases'], LIBPATH=['lib_qp'], CPPPATH=cpp_path)
 env.SharedLibrary('mpc2', mpc_files, LIBS=['m', 'qpoases'], LIBPATH=['lib_qp'], CPPPATH=cpp_path)
