+++ selfdrive/controls/lib/lead_mpc_lib/lib_mpc_export/acado_auxiliary_functions.h
